Separate reporters to be simpler to plugin
