+++ dev/public/index.html
@@ -93,5 +93,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/rickshaw/1.5.1/rickshaw.min.js"></script>
     <script src="/js/nes.js"></script>
     <script src="/js/app.js"></script>
+    <script src="/js/reporters/memory.js"></script>
+    <script src="/js/reporters/register.js"></script>
   </body>
 </html>
