+++ lib/routes.js
@@ -24,8 +24,6 @@ module.exports = [
         };
 
         const source = (info && (info.hostname + '-' + info.pid)) || 'unknown';
-
-
         this.report(source, 'register', info);
 
         reply({ clientId: request.connection.clientId });
@@ -42,9 +40,10 @@ module.exports = [
         const client = registered[request.auth.credentials.clientId] || {};
         const source = (client.info && (client.info.hostname + '-' + client.info.pid)) || 'unknown';
 
-        if (payload.memory) {
-          this.report(source, 'memory', payload.memory);
-        }
+        const types = Object.keys(payload);
+        types.forEach((type) => {
+          this.report(source, type, payload[type]);
+        });
 
         reply();
       }
