+++ package.json
@@ -14,7 +14,7 @@
   },
   "license": "MIT",
   "scripts": {
-    "start": "node dev/server.js & open http://localhost:5000/",
+    "start": "node dev/server.js",
     "test": "belly-button && lab"
   },
   "engines": {
